Removed fancy error pages (Sorry!)
I have to remove this feature, sorry! But, I have a limited dev schedule, and would you rather have a half-implemented, buggy feature that is easy to configure, or a stable feature that has little/no configuration options? And. I don't want to delay this project more than I have to.